+++ changelog.md
@@ -1,5 +1,6 @@
 * 4.1.0
    - Added pass SASL version to kpro_auth_backend behaviour modules
+   - The application ˋstartˋ method must return the `pid` of the top supervisor
 * 4.0.3
    - Fix type spec for `hostname()` to include `inet:ip_address()`
 * 4.0.2